fix: add Stellar address validation to scheduled payments
- Add validateStellarAddress check in handleCreate before posting
- Extract validateStellarAddress to shared frontend/src/utils/validation.js
- Update BatchPayment.jsx to import from shared utility (dedup)
- Include required execute_at field in scheduled payment POST body
- Clean up unused Edit2 import
