+++ frontend/src/pages/ScheduledPayments.jsx
@@ -1,11 +1,12 @@
 import React, { useState, useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { ArrowLeft, Plus, Trash2, Edit2, List, Calendar } from 'lucide-react';
-import api from '../utils/api';
+import { ArrowLeft, Plus, Trash2, List, Calendar } from 'lucide-react';
 import toast from 'react-hot-toast';
 import { useTranslation } from 'react-i18next';
+import api from '../utils/api';
 import { CURRENCIES } from '../utils/currency';
 import ScheduledPaymentsCalendar from '../components/ScheduledPaymentsCalendar';
+import { validateStellarAddress } from '../utils/validation';
 
 export default function ScheduledPayments() {
   const navigate = useNavigate();
@@ -39,14 +40,23 @@ export default function ScheduledPayments() {
 
   const handleCreate = async (e) => {
     e.preventDefault();
-    if (!form.recipient_wallet || !form.amount) {
-      toast.error(t('scheduled.invalid') || 'Invalid input');
+
+    // Validate Stellar address format
+    const addressError = validateStellarAddress(form.recipient_wallet);
+    if (addressError) {
+      toast.error(addressError);
+      return;
+    }
+
+    if (!form.amount || parseFloat(form.amount) <= 0) {
+      toast.error(t('scheduled.invalid') || 'Please enter a valid amount');
       return;
     }
 
     try {
       await api.post('/scheduled-payments', {
-        recipient_wallet: form.recipient_wallet,
+        execute_at: new Date(Date.now() + 3600000).toISOString(),
+        recipient_wallet: form.recipient_wallet.trim(),
         amount: parseFloat(form.amount),
         asset: form.asset,
         frequency: form.frequency,
